Nytt schema og egen form for bedkom
